agents_sim_tidy Example output from SIR simulations in a data.frame format

This simulated data is an example of the output from the function simulate_SIR_agents() when output_format = "data.frame". It is a 9400 by 5 data.frame where each row corresponds to the agents statistic of a given agent and simulation. Each row is asssociated with 1 agent and contains is a "sufficient" statistic for each agent's infection. Each agent's infection is uniquely identified by an initial state, max time before infection (or time T-1), and max time before recovery (or time T-1). Note time T-1 is the final recorded time in the SIR model (in this case T = 50).

agents_sims_tidy

Format

An 9400 x 5 data frame where each row is an agent for a given simulation and the columns are

sim

Simulation number between 1 and 50

individual

This is the ID of the individual

init_state

The initial state of the agent at time 0. It is 0/1/2 for susceptible, infectious, recovered, respectively.

SMax

maximum time individual was suspectable (S)

IMax

maximum time individual was suspectable (I)

Examples

## show first 10 simulations of the first individual
agents_sims_tidy[1:10,]

#> # A tibble: 10 x 5
#>    sim   individual init_state  SMax  IMax
#>    <fct> <fct>           <dbl> <dbl> <dbl>
#>  1 1     1                   0    NA    NA
#>  2 2     1                   0    22    27
#>  3 3     1                   0    NA    NA
#>  4 4     1                   0    15    20
#>  5 5     1                   0    NA    NA
#>  6 6     1                   0    NA    NA
#>  7 7     1                   0    NA    NA
#>  8 8     1                   0    NA    NA
#>  9 9     1                   0    NA    NA
#> 10 10    1                   0    15    21