Fortify agent data frame with columns when individual stops being susceptible and stops being infected (as well as initial state).

Fortify agent data frame with columns when individual stops being susceptible and stops being infected (as well as initial state).

fortify_agents(
  data,
  time_col = c("tI", "tR"),
  max_time = floor(max(raw_df[, time_col], na.rm = TRUE)) + 1
)

Arguments

data
data frame, agent based data frame, class ("individuals_df")

time_col
length 2 string vector, column names recording when individual is infected and when they enter the recovery stage

max_time
int, maximum time for infection process

Value

fortified_df data frame, the raw_df plus three additional columns:

init_state

Initial state for individual (at time t = 0). For the states, 0 = S, 1 = I, 2 = R.

max_time_S

maximum time individual was susceptible (S)

max_time_I

maximum time individual was infected (I)

Examples

fortify_df <- fortify_agents(timeternR::hagelloch_raw,
                             time_col = c("tI","tR"),
                             max_time = 95)
assertthat::are_equal(fortify_df[,c("init_state",
"max_time_S", "max_time_I")],
                      timeternR::hagelloch_agents)

#> [1] TRUE